+++ app/src/test/java/com/weighttrack/data/sync/SyncStoreTest.kt
@@ -49,8 +49,8 @@ class SyncStoreTest {
     fun setUp() {
         phone = database()
         tablet = database()
-        phoneStore = SyncStore(phone.syncDao(), phone.deletionDao())
-        tabletStore = SyncStore(tablet.syncDao(), tablet.deletionDao())
+        phoneStore = SyncStore(phone, phone.syncDao(), phone.deletionDao())
+        tabletStore = SyncStore(tablet, tablet.syncDao(), tablet.deletionDao())
     }
 
     @After